requirements.txt
torch
transformers
sentence-transformers
faiss-cpu
langchain
pdfminer.six
pytesseract
Pillow
huggingface-hub
